# perf(ext/headers): optimize getHeader using for loop